feat: rewritten the whole widget logic and updated the weather and media widget
The widget logic was changed where it divided the icon and label styles different wherever it applies.

The media widget now fetches the thumbnail and now does not crashes randomly (hopefully)

The weather widget has also been updated with a new provider weatherapi.

The new update also has a lot of config changes and may require to rewrite certain sections.

Co-Authored-By: amnweb <16545063+amnweb@users.noreply.github.com>